# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  controllers.NewType:
    properties:
      typeName:
        example: Serial
        minLength: 2
        type: string
    required:
    - typeName
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API Server for Ozinwe
  title: Ozinwe API
  version: "1.0"
paths:
...
  /admin/type/{id}/delete:
    delete:
      consumes:
      - application/json
      operationId: delete-type
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteTypeOfProject
      tags:
      - admin-movie-type-controller
  /admin/type/{id}/edit:
    get:
      consumes:
      - application/json
      operationId: edit-type
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: EditTypeOfProject
      tags:
      - admin-movie-type-controller
  /admin/type/{id}/update:
    put:
      consumes:
      - application/json
      operationId: update-type
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: typeName
        in: body
        name: typeName
        required: true
        schema:
          $ref: '#/definitions/controllers.NewType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateTypeOfProject
      tags:
      - admin-movie-type-controller
  /admin/type/create:
    post:
      consumes:
      - application/json
      operationId: create-type
      parameters:
      - description: typeName
        in: body
        name: typeName
        required: true
        schema:
          $ref: '#/definitions/controllers.NewType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateTypeOfProject
      tags:
      - admin-movie-type-controller
...
  /movie/{id}/favorite:
    delete:
      consumes:
      - application/json
      operationId: delete-movie-from-favorite
      parameters:
      - description: movieID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteMovieFromFavorite
      tags:
      - movie-controller
    post:
      consumes:
      - application/json
      operationId: add-movie-to-favorite
      parameters:
      - description: movieID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: AddMovieToFavorite
      tags:
      - movie-controller
...
  /movie/favorite:
    get:
      consumes:
      - application/json
      operationId: get-all-favorite-movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetAllFavoriteMovies
      tags:
      - movie-controller
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"